+++ linux/keyman-config/keyman_config/__init__.py
@@ -21,8 +21,8 @@ def _(txt):
 gettext.bindtextdomain('keyman-config', '/usr/share/locale')
 gettext.textdomain('keyman-config')
 
-if __tier__ == 'alpha':
-    # Alpha versions will work against the staging server so that they
+if __tier__ == 'alpha' or __tier__ == 'beta':
+    # Alpha and beta versions will work against the staging server so that they
     # can access new APIs etc that will only be available there. The staging
     # servers have resource constraints but should be okay for limited use.
     KeymanComUrl = 'https://keyman-staging.com'