+++ windows/src/developer/TIKE/project/Keyman.Developer.UI.Project.UfrmNewProjectParameters.pas
@@ -476,7 +476,7 @@ begin
     Tag := StrToken(Tags, ' ');
     BCP47Tag := TBCP47Tag.Create(Tag);
     try
-      BCP47Tag.Tag := TCanonicalLanguageCodeUtils.FindBestTag(BCP47Tag.Tag, False);
+      BCP47Tag.Tag := TCanonicalLanguageCodeUtils.FindBestTag(BCP47Tag.Tag, False, False);
       if BCP47Tag.IsValid(False) then
       begin
         Language := TPackageKeyboardLanguage.Create(pack);